fix RecipePageFragment layout 2
