+++ app/src/main/java/com/itis/summerproject23/UserAccountFragment.kt
@@ -4,11 +4,8 @@ import android.os.Bundle
 import android.view.View
 import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.lifecycleScope
-import com.itis.summerproject23.database.UsersDatabase
-import kotlinx.coroutines.launch
 
-class UserAcoountFragment : Fragment(R.layout.fragment_user_account) {
+class UserAccountFragment : Fragment(R.layout.fragment_user_account) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
